requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
build==1.2.1
    # via pip-tools
click==8.1.7
    # via pip-tools
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.52.4
    # via matplotlib
kiwisolver==1.4.5
    # via matplotlib
matplotlib==3.9.0
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   pandas
opencv-python==4.9.0.80
    # via -r requirements.in
packaging==24.0
    # via
    #   build
    #   matplotlib
pandas==2.2.2
    # via -r requirements.in
pillow==10.3.0
    # via
    #   -r requirements.in
    #   matplotlib
pip-tools==7.4.1
    # via -r requirements.in
pyparsing==3.1.2
    # via matplotlib
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pyserial==3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via
    #   build
    #   pip-tools
tzdata==2024.1
    # via pandas
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
